Refactor tests

+++ photo/tests/test_s3_client/test_s3_client.py
@@ -1,4 +1,3 @@
-import botocore
 from django.test import TestCase
 
 from integrations.aws.s3 import Client
@@ -9,15 +8,29 @@ class S3BucketTest(TestCase):
         self.client = Client()
         return super().setUp()
 
-    def test_s3(self):
+    def test_create(self):
         response_create = self.client.create_bucket(bucket_name="test")
         self.assertEqual(response_create["ResponseMetadata"]["HTTPStatusCode"], 200)
 
-        response_head = self.client.head_bucket(bucket_name="test")
-        self.assertEqual(response_head["ResponseMetadata"]["HTTPStatusCode"], 200)
+    def test_bucket_exists(self):
+        self.client.create_bucket(bucket_name="test-existence")
+        exists = self.client.check_bucket_existence(bucket_name="test-existence")
+        not_exists = self.client.check_bucket_existence(bucket_name="not_existence")
 
-        response_delete = self.client.delete_bucket(bucket_name="test")
+        self.assertEqual(exists, True)
+        self.assertEqual(not_exists, False)
+
+    def test_bucket_delete(self):
+        self.client.create_bucket(bucket_name="test-delete")
+        response_delete = self.client.delete_bucket(bucket_name="test-delete")
+        bucket_exists = self.client.check_bucket_existence(bucket_name="test-delete")
+
         self.assertEqual(response_delete["ResponseMetadata"]["HTTPStatusCode"], 204)
+        self.assertEqual(bucket_exists, False)
 
-        with self.assertRaises(botocore.exceptions.ClientError):
-            self.client.head_bucket(bucket_name="test")
+    def tearDown(self) -> None:
+        if self.client.check_bucket_existence("test"):
+            self.client.delete_bucket(bucket_name="test")
+        if self.client.check_bucket_existence("test-existence"):
+            self.client.delete_bucket(bucket_name="test-existence")
+        return super().tearDown()
